requirements.txt
affine==2.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fiona==1.9.5
fonttools==4.49.0
fsspec==2024.2.0
geojson==3.1.0
geopandas==0.13.2
graphviz==0.20.1
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
ipdb==0.13.13
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest_asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
platformdirs==4.2.0
prompt-toolkit==3.0.42
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
pystac==1.9.0
pystac-client==0.7.6
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
scipy==1.13.0
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
snuggs==1.4.7
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
tifffile==2024.2.12
tomli==2.0.1
torch==2.2.1
torchaudio==2.2.1
torchsummary==1.5.1
torchvision==0.17.1
torchviz==0.0.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
up42-py==0.36.0
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.42.0
zipp==3.17.0
